+++ cognite/client/_api/sequences.py
@@ -763,7 +763,7 @@ class SequencesAPI(APIClient):
             resource_cls=Sequence,
             method="POST",
             limit=limit,
-            advanced_filter=filter.dump(camel_case=True) if isinstance(filter, Filter) else filter,
+            advanced_filter=filter.dump(camel_case_property=True) if isinstance(filter, Filter) else filter,
             sort=prepare_filter_sort(sort, SequenceSort),
             api_subversion="beta",
         )