+++ cognite/client/_api_client.py
@@ -452,7 +452,7 @@ class APIClient:
                         body["filter"] = filter
                     if advanced_filter:
                         body["advancedFilter"] = (
-                            advanced_filter.dump(camel_case=True)
+                            advanced_filter.dump(camel_case_property=True)
                             if isinstance(advanced_filter, Filter)
                             else advanced_filter
                         )
@@ -603,7 +603,7 @@ class APIClient:
                     }
                     if advanced_filter:
                         body["advancedFilter"] = (
-                            advanced_filter.dump(camel_case=True)
+                            advanced_filter.dump(camel_case_property=True)
                             if isinstance(advanced_filter, Filter)
                             else advanced_filter
                         )