+++ tests/tests_integration/test_api/test_relationships.py
@@ -200,6 +200,16 @@ class TestRelationshipscognite_client:
         assert res[0].source == asset
         assert res[0].target == time_series
 
+    def test_retrieve_relationship_with_resource_client_set(
+        self, cognite_client: CogniteClient, relationship_with_resources
+    ) -> None:
+        relationship, ext_id, asset, time_series = relationship_with_resources
+
+        res = cognite_client.relationships.retrieve(ext_id, fetch_resources=True)
+
+        assert res.source._get_cognite_client() is not None
+        assert res.target._get_cognite_client() is not None
+
     def test_retrieve_unknown_raises_error(self, cognite_client: CogniteClient):
         with pytest.raises(CogniteNotFoundError) as e:
             cognite_client.relationships.retrieve_multiple(external_ids=["this does not exist"])