+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Changes are grouped as follows
 ### Fixed
 - Bug when `cognite.client.data_classes.filter` used with any `data_modeling` endpoint raised a `CogniteAPIError` for 
   snake_cased properties. This is now fixed. 
+- When calling `client.relationships.retrieve`, `.retrieve_multiple`, or `.list` with `fetch_resources=True`, the 
+  `target` and `source` resources were not instantiated with a `cognite_client`. This is now fixed.
 
 ## [7.0.2] - 2023-11-15
 ### Fixed